change sqrt to sqrtf to make sure single-precision calculation.\n\nIt seems sqrt(float) emits sqrtsd on xcode.

+++ include/kismet/ai/fuzzy/fuzzy_hedge_fairly.h
@@ -28,7 +28,7 @@ public:
 
     void aggregate(float dom) override
     {
-        m_term->aggregate(std::sqrt(dom));
+        m_term->aggregate(std::sqrtf(dom));
     }
 private:
     fuzzy_term_ptr m_term;
